Added messeges.properties & modified DBConnect.java file. Restructured folder Structure. Added resources in gitignore.
