+++ src/main/java/com/db/DBConnect.java
@@ -3,6 +3,7 @@ package com.db;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
+import java.util.ResourceBundle;
 
 public class DBConnect {
 
@@ -10,7 +11,11 @@ public class DBConnect {
 		Connection con=null;
 			try {
 				//Class.forName("com.mysql.jdbc.Driver");
-				con=DriverManager.getConnection("jdbc:mysql://localhost:3306/MedManagement","root","0811");
+				ResourceBundle message = ResourceBundle.getBundle("messages");
+				System.out.println(message.getString("SQL_URL"));
+				System.out.println(message.getString("SQL_USER"));
+				System.out.println(message.getString("SQL_PASSWORD"));
+				con=DriverManager.getConnection(message.getString("SQL_URL"),message.getString("SQL_USER"),message.getString("SQL_PASSWORD"));
 			} catch (SQLException e) {
 				e.printStackTrace();
 			}
