+++ src/main/resources/messages.properties
@@ -1,0 +1,3 @@
+SQL_URL = jdbc:mysql://localhost:3306/MedManagement
+SQL_USER = root
+SQL_PASSWORD = root
